+++ src/main/java/kitchenpos/dao/MenuRepositoryImpl.java
@@ -12,7 +12,7 @@ import kitchenpos.dao.mapper.ProductMapper;
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuGroup;
 import kitchenpos.domain.MenuProduct;
-import kitchenpos.domain.MenuRepository;
+import kitchenpos.domain.repository.MenuRepository;
 import kitchenpos.domain.Product;
 import org.springframework.stereotype.Repository;
