+++ src/main/java/kitchenpos/domain/repository/OrderRepository.java
@@ -1,7 +1,8 @@
-package kitchenpos.domain;
+package kitchenpos.domain.repository;
 
 import java.util.List;
 import java.util.Optional;
+import kitchenpos.domain.Order;
 
 public interface OrderRepository {
