+++ src/main/java/kitchenpos/domain/repository/TableGroupRepository.java
@@ -1,7 +1,8 @@
-package kitchenpos.domain;
+package kitchenpos.domain.repository;
 
 import java.util.List;
 import java.util.Optional;
+import kitchenpos.domain.TableGroup;
 
 public interface TableGroupRepository {
