+++ src/test/java/kitchenpos/application/MenuGroupServiceTest.java
@@ -5,14 +5,13 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.util.List;
 import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.MenuGroupRepository;
+import kitchenpos.domain.repository.MenuGroupRepository;
 import kitchenpos.fixture.MenuGroupFixture;
 import kitchenpos.support.ServiceIntegrationTest;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
 
 class MenuGroupServiceTest extends ServiceIntegrationTest {
